requirements.txt
# ==============================================================================
#                       AirGorithm Project Requirements
# ==============================================================================

# ------------------------------------------------------------------------------
#                       Core Dependencies
# ------------------------------------------------------------------------------
numpy==1.23.5 # Numerical computing library
pandas        # Data analysis and manipulation library
scikit-learn  # Machine learning algorithms and tools
scipy         # Scientific computing tools

# ------------------------------------------------------------------------------
#                       TensorFlow and Related Libraries
# ------------------------------------------------------------------------------
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.15.0; sys_platform == 'darwin' and 'ARM' in platform_version  # Mac avec puces M1
tensorflow==2.15.0; sys_platform == 'darwin' and 'ARM' not in platform_version  # Mac avec processeurs Intel
tensorflow==2.15.0; sys_platform != 'darwin'  # Windows & Ubuntu
h5py          # HDF5 file format library (required by TensorFlow)

# ------------------------------------------------------------------------------
#                       PyTorch and Related Libraries
# ------------------------------------------------------------------------------
torch  # PyTorch
torchvision  # Vision library
torchaudio  # Audio library

# ------------------------------------------------------------------------------
#                       Model Tools
# ------------------------------------------------------------------------------
ultralytics==8.3.93  # YOLOv8 framework
onnx         # ONNX format support
onnxruntime  # ONNX inference optimization

# ------------------------------------------------------------------------------
#                       Roboflow Integration
# ------------------------------------------------------------------------------
roboflow     # Roboflow Python package

# ------------------------------------------------------------------------------
#                       Data Processing & Visualization
# ------------------------------------------------------------------------------
opencv-python  # OpenCV for image processing
pillow        # Pillow for image handling
matplotlib    # Plotting library
seaborn       # Statistical data visualization
shapely       # Geometry operations (used for bounding boxes)
imgaug        # Image augmentation library
albumentations  # Advanced image augmentation
scikit-image  # Additional image processing tools

# ------------------------------------------------------------------------------
#                       GPU Optimization (for A100)
# ------------------------------------------------------------------------------
nvidia-pyindex  # NVIDIA package index
nvidia-tensorRT  # TensorRT for inference optimization
apex          # NVIDIA Apex for mixed precision training
cupy-cuda11x  # GPU-accelerated NumPy operations
xformers      # Memory-efficient attention computation (useful for YOLO models)

# ------------------------------------------------------------------------------
#                       Google Cloud Platform (GCP)
# ------------------------------------------------------------------------------
google-cloud-bigquery  # Google BigQuery integration
google-cloud-storage  # Google Cloud Storage integration
google-api-core       # Google API core library
googleapis-common-protos  # Common protobufs for Google APIs
protobuf        # Protocol Buffers
db-dtypes       # Database data types for pandas
pyarrow         # Arrow columnar data format

# ------------------------------------------------------------------------------
#                       MLflow and Prefect
# ------------------------------------------------------------------------------
mlflow        # MLflow for experiment tracking
prefect       # Prefect for workflow orchestration

# ------------------------------------------------------------------------------
#                       API Dependencies
# ------------------------------------------------------------------------------
fastapi       # FastAPI for building APIs
pytz          # Timezone support
uvicorn       # ASGI server for running FastAPI
requests   # Bibliothèque HTTP pour Python, permettant d'envoyer facilement des requêtes HTTP et de gérer les réponses.

# ------------------------------------------------------------------------------
#                       Development & Testing Tools
# ------------------------------------------------------------------------------
jupyterlab    # JupyterLab IDE
ipykernel     # IPython kernel for Jupyter
jupyter       # Jupyter Notebook
pytest        # Testing framework
pylint        # Code linting tool
ipdb          # Interactive Python debugger
ipython       # Interactive Python shell
python-dotenv # Loads environment variables from .env file
psycopg2-binary  # PostgreSQL adapter for Python
psycopg2  # PostgreSQL adapter for Python

# ------------------------------------------------------------------------------
#                       Testing
# ------------------------------------------------------------------------------
httpx        # HTTP client for testing
pytest-asyncio  # Asyncio support for pytest

# ------------------------------------------------------------------------------
#                       Utilities
# ------------------------------------------------------------------------------
tqdm         # Progress bar library

# ------------------------------------------------------------------------------
#                       Interface Utilisateur (UI) pour le Déploiement
# ------------------------------------------------------------------------------
streamlit    # Interface simple pour tester l’API YOLO
gradio       # Interface interactive pour tester le modèle YOLO
dash         # Web framework pour visualisation avancée

# ------------------------------------------------------------------------------
#                       Unnecessary Dependencies (Commented Out)
# ------------------------------------------------------------------------------
# wheel      # Package distribution format (usually handled by pip)
# nbresult   # Unclear purpose, consider removing if not used
# colorama   # Unclear purpose, consider removing if not used
# yapf       # Unclear purpose, consider removing if not used
